+++ src/rules/no_default_default.rs
@@ -36,3 +36,44 @@ pub fn no_default_default_rule() -> Rule {
         }
     }
 }
+
+#[cfg(test)]
+mod tests {
+    use proc_macros::rule_tests;
+
+    use super::*;
+    use crate::RuleTester;
+
+    #[test]
+    fn test_no_default_default_rule() {
+        const ERROR_MESSAGE: &str = "Use '_d()' instead of 'Default::default()'";
+
+        RuleTester::run(
+            no_default_default_rule(),
+            rule_tests! {
+                valid => [
+                    r#"
+                        fn foo() {
+                            let bar = Default::something_else::default();
+                        }
+                    "#,
+                ],
+                invalid => [
+                    {
+                        code => r#"
+                            fn foo() {
+                                let bar = Default::default();
+                            }
+                        "#,
+                        errors => [ERROR_MESSAGE],
+                        output => r#"
+                            fn foo() {
+                                let bar = _d();
+                            }
+                        "#,
+                    },
+                ]
+            },
+        );
+    }
+}
